requirements.txt
# Core
python-dotenv==1.0.0

# LangChain
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-chroma==0.1.4

# Vector Store
chromadb==0.5.0

# LLM API
openai>=1.12.0

# Monitoring
langsmith>=0.1.17
langfuse>=2.0.0

# Interface
gradio==6.0.0

# Utils
requests>=2.31.0
pandas>=2.0.0
pydantic>=2.7.4,<3.0.0

# Development
pytest>=7.4.0

# Free embeddings alternative (HuggingFace)
sentence-transformers>=2.2.0

# PDF Support
pypdf>=3.17.0
